+++ backend/src/main/java/com/techeer/backend/api/resume/dto/response/ResumePageResponse.java
@@ -1,7 +1,7 @@
-package com.techeer.backend.domain.resume.dto.response;
+package com.techeer.backend.api.resume.dto.response;
 
 
-import com.techeer.backend.domain.resume.entity.Resume;
+import com.techeer.backend.api.resume.domain.Resume;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -17,6 +17,9 @@ public class ResumePageResponse {
     private int currentPage;
 
     private ResumePageResponse(List<ResumePageElement> elementList, int totalPages, int number) {
+        this.elementList = elementList;
+        this.totalPage = totalPages;
+        this.currentPage = number;
     }
 
     public static ResumePageResponse from(List<ResumePageElement> elementList , Page<Resume> page) {
